add kr champ rank data again
